hw-14 * Перенес работу с centrifugo в отдельный хук. Тем самым починил работу centrifugo после выхода и входа в профиль (ранее подписка не обновлялась и после перезахода события centrifugo не приходили) * Починил отправку уведомлений о новых сообщений (из-за замыкания, после перезахода список чатов не обновлялся, в функции, которая шлет уведомления)
